# server/controllers/reservations.controller.js
exports.listReservations = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.createReservation = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.getReservation = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.updateReservation = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.cancelReservation = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};