# server/controllers/rooms.controller.js
exports.listRooms = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.createRoom = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.getRoomDetails = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.updateRoom = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};

exports.deleteRoom = (req, res) => {
    res.status(501).json({ message: "Endpoint não implementado." });
};